Bug 970271: Fix GCC warnings about comparison of signed and unsigned values, r=rjesup
UltraBlame original commit: 26d23e6ded3ae273736cfc333040d719e3664738

+++ netwerk/protocol/rtsp/controller/RtspControllerChild.cpp
@@ -123,7 +123,7 @@ RtspControllerChild::RecvOnConnected(
   AddMetaData(meta.forget());
 
   
-  if ((index + 1) == mTotalTracks) {
+  if ((static_cast<uint32_t>(index) + 1) == mTotalTracks) {
     
     if (mListener) {
       mListener->OnConnected(index, nullptr);
